Truck Networking - fetchAllTrucks
