+++ FoodTruckTracker/.swiftlint.yml
@@ -71,7 +71,7 @@ file_length:
   warning: 800
   error: 1000
 line_length:
-  warning: 200
+  warning: 220
   ignores_comments: true
   ignores_interpolated_strings: true
 multiline_arguments:
