+++ FoodTruckTracker/FoodTruckTracker/Helpers/Constants.swift
@@ -26,3 +26,23 @@ enum NetworkError: Error {
     case noRep
     case otherError
 }
+
+enum UserType {
+    case diner
+    case owner
+}
+
+enum Cuisine: String, CaseIterable {
+    case other = "Other"
+    case african = "African"
+    case american = "American"
+    case asian = "Asian"
+    case cuban = "Cuban"
+    case european = "European"
+    case mexican = "Mexican"
+    case middleEastern = "Middle Eastern"
+    case southAmerican = "South American"
+    case bakery = "Bakery"
+    case breakfast = "Breakfast"
+    case treats = "Treats"
+}
