--- FoodTruckTracker/FoodTruckTracker/ModelControllers/TruckController.swift
@@ -1,34 +1,0 @@
-//
-//  TruckController.swift
-//  FoodTruckTracker
-//
-//  Created by Cora Jacobson on 10/14/20.
-//
-
-import Foundation
-import UIKit
-
-class TruckController {
-    
-    func fetchImage(at urlString: String, completion: @escaping (Result<UIImage, NetworkError>) -> Void) {
-        let imageURL = URL(string: urlString)!
-        var request = URLRequest(url: imageURL)
-        request.httpMethod = HTTPMethod.get.rawValue
-        let task = URLSession.shared.dataTask(with: request) { data, _, error in
-            if let error = error {
-                print("Error receiving image: \(urlString), error: \(error)")
-                completion(.failure(.tryAgain))
-                return
-            }
-            guard let data = data else {
-                print("No data received from fetchImage")
-                completion(.failure(.noData))
-                return
-            }
-            let image = UIImage(data: data)!
-                completion(.success(image))
-        }
-        task.resume()
-        }
-    
-}
