+++ FoodTruckTracker/FoodTruckTracker/Models/TruckListing.swift
@@ -21,8 +21,20 @@ struct TruckListing: Codable {
     var identifier: Int
     var name: String
     var location: String
-    var departureTime: Date
-    var cuisine: Cuisine.RawValue
+    var departureTime: Date = Date() + (4 * 60 * 60)
+    var cuisine: String
     var imageString: String
-    var menu: [MenuItem]
+    var menu: [MenuItem] = []
+    var ratings: [Int] = []
+    
+    init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: Keys.self)
+        identifier = try container.decode(Int.self, forKey: .id)
+        name = try container.decode(String.self, forKey: .name)
+        location = try container.decode(String.self, forKey: .location)
+//        departureTime = try container.decode(Date.self, forKey: .departureTime)
+        let cuisineId = try container.decode(Int.self, forKey: .cuisineId)
+        cuisine = Cuisine.allCases[cuisineId].rawValue
+        imageString = try container.decode(String.self, forKey: .photoUrl)
+    }
 }
